+++ index.js
@@ -125,7 +125,7 @@ const log = function (input, color = '#aaa', style = 'font-weight:bold') {
 /**
  * @function hasEngine
  * @param {Engine} engine
- * @param {string} version
+ * @param {string=} version
  * @returns {boolean}
  */
 export function hasEngine(engine, version) {
@@ -138,7 +138,7 @@ export function hasEngine(engine, version) {
 /**
  * @function engineExists
  * @param {Engine} engine
- * @param {string} version
+ * @param {string=} version
  * @returns {boolean}
  */
 export function engineExists(engine, version = null) {
@@ -187,6 +187,7 @@ export function setOptions(options) {
  * @async
  * @function setEngine
  * @param {Engine} engine
+ * @param {string=} version
  */
 export async function setEngine(engine, version = null) {
   if (!hasEngine(engine, version)) {
@@ -384,6 +385,7 @@ export function interpretErrorMessage(error, code, engine) {
  * @async
  * @function loadEngine
  * @param {Engine} engine
+ * @param {{useEngine?: Engine, version?: string}=} userOptions
  * @returns {boolean} If engine was loaded (or already loaded)
  * @throws {Error} If engine does not exist
  */
@@ -1072,7 +1074,7 @@ export async function loadEngines(engines) {
 /**
  * @async
  * @function runCode
- * @param {{use?: Engine}=} userOptions
+ * @param {{use?: Engine, version?: string}=} userOptions
  * @returns {any=} Last result from pyodide. (Not the other runners)
  * @throws {Error|PythonError} Invalid python engine | A python error
  */
@@ -1092,7 +1094,7 @@ export async function runCode(code, userOptions = {}) {
 /**
  * @async
  * @function getVariable
- * @param {{use?: Engine}=} userOptions
+ * @param {{use?: Engine, version?: string}=} userOptions
  * @returns {any}
  */
 export async function getVariable(name, userOptions = {}) {
@@ -1108,7 +1110,7 @@ export async function getVariable(name, userOptions = {}) {
 /**
  * @async
  * @function getVariables
- * @param {{use?: Engine, includeValues?: boolean, filter?: null | (name) => boolean, onlyShowNewVariables?: boolean}=} userOptions
+ * @param {{use?: Engine, version?: string, includeValues?: boolean, filter?: null | (name) => boolean, onlyShowNewVariables?: boolean}=} userOptions
  * @returns {Variables|string[]}
  */
 export async function getVariables(
@@ -1141,10 +1143,11 @@ export async function getVariables(
  * @function setVariable
  * @param {string} name
  * @param {any} value
- * @param {{use?: Engine}=} userOptions
+ * @param {{use?: Engine, version?: string}=} userOptions
  */
 export async function setVariable(name, value, userOptions = {}) {
-  const { use: specificEngine = pythonRunner.currentEngine } = userOptions;
+  const { use: specificEngine = pythonRunner.currentEngine, version = null } =
+    userOptions;
 
   if (!hasEngine(specificEngine, version))
     await loadEngine(specificEngine, { version });
@@ -1159,10 +1162,11 @@ export async function setVariable(name, value, userOptions = {}) {
  * @async
  * @function setVariables
  * @param {Variables} variables
- * @param {{use?: Engine}=} userOptions
+ * @param {{use?: Engine, version?: string}=} userOptions
  */
 export async function setVariables(variables, userOptions = {}) {
-  const { use: specificEngine = pythonRunner.currentEngine } = userOptions;
+  const { use: specificEngine = pythonRunner.currentEngine, version = null } =
+    userOptions;
 
   if (!hasEngine(specificEngine, version))
     await loadEngine(specificEngine, { version });
@@ -1176,10 +1180,11 @@ export async function setVariables(variables, userOptions = {}) {
  * @async
  * @function clearVariable
  * @param {string} name
- * @param {{use?: Engine}=} userOptions
+ * @param {{use?: Engine, version?: string}=} userOptions
  */
 export async function clearVariable(name, userOptions = {}) {
-  const { use: specificEngine = pythonRunner.currentEngine } = userOptions;
+  const { use: specificEngine = pythonRunner.currentEngine, version = null } =
+    userOptions;
 
   if (!hasEngine(specificEngine, version))
     await loadEngine(specificEngine, { version });
@@ -1190,10 +1195,11 @@ export async function clearVariable(name, userOptions = {}) {
 /**
  * @async
  * @function clearVariables
- * @param {{use?: Engine}=} userOptions
+ * @param {{use?: Engine, version?: string}=} userOptions
  */
 export async function clearVariables(userOptions = {}) {
-  const { use: specificEngine = pythonRunner.currentEngine } = userOptions;
+  const { use: specificEngine = pythonRunner.currentEngine, version = null } =
+    userOptions;
 
   if (!hasEngine(specificEngine, version))
     await loadEngine(specificEngine, { version });